Move Separate component from UI kit to Onboarding app
