+++ apps/onboarding/src/components/language-switcher.tsx
@@ -1,7 +1,8 @@
 import styled from '@emotion/styled'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-import { Separate, theme } from 'ui'
+import { Children, Fragment } from 'react'
+import { theme } from 'ui'
 import { useCurrentMarket } from '@/lib/l10n'
 
 const FALLBACK_PATH = '/'
@@ -27,6 +28,38 @@ export const LanguageSwitcher = () => {
   )
 }
 
+const StyledLink = styled(Link)<{ active: boolean }>(({ active }) => ({
+  textDecoration: 'none',
+  '&:hover': {
+    color: theme.colors.gray900,
+  },
+  color: active ? theme.colors.gray900 : theme.colors.gray500,
+}))
+
+type SeparateProps = {
+  children: React.ReactNode
+  Separator: React.ReactNode
+  className?: string
+  as?: React.ElementType
+}
+
+const Separate = ({ children, Separator, className, as: Element = 'div' }: SeparateProps) => {
+  const array = Children.toArray(children)
+
+  const childrenWithSeparators = array.map((childItem, index) => (
+    <Fragment key={index}>
+      {index > 0 && Separator}
+      {childItem}
+    </Fragment>
+  ))
+
+  if (className) {
+    return <Element className={className}>{childrenWithSeparators}</Element>
+  }
+
+  return <>{childrenWithSeparators}</>
+}
+
 const Wrapper = styled(Separate)({
   display: 'flex',
   height: '1.5rem',
@@ -39,11 +72,3 @@ const Separator = styled.div({
   marginLeft: '0.5rem',
   marginRight: '0.5rem',
 })
-
-const StyledLink = styled(Link)<{ active: boolean }>(({ active }) => ({
-  textDecoration: 'none',
-  '&:hover': {
-    color: theme.colors.gray900,
-  },
-  color: active ? theme.colors.gray900 : theme.colors.gray500,
-}))
